requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==5.0.4
django-extensions==3.2.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-seed==0.3.1
executing==2.0.1
Faker==25.0.0
ipython==8.24.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
toposort==1.10
traitlets==5.14.3
typing_extensions==4.11.0
wcwidth==0.2.13
